refactored cache
