+++ src/index.js
@@ -1,5 +1,4 @@
 const express = require('express')
-const redis = require('redis')
 const constants = require('./config/constants')
 const logger = require('./util/logger')
 const router = require('./router')
@@ -14,15 +13,3 @@ app.use('', router)
 app.listen(constants.port, () => {
     logger.i(`Server started at port ${constants.port}`)
 })
-
-const clientRedis = redis.createClient(constants.redisUrl)
-
-clientRedis.on('connect', () => {
-    logger.i(`Server connect with redis`)
-});
-
-clientRedis.on('error', err => {
-    logger.i(`Server redis with error: ${err}`)
-});
-
-module.exports.clientRedis  = clientRedis
